feat: ✨ error handing
